Add files via upload
Implementation of  an ideal ADC in Verilog-A. The internal resolution of the ADC will be 8 bit. The ADC should have
three ports: one CLK input, the analog input and a 8 bit wide output port.  Since the ADC is written in Verilog-A, you have to write this part in Cadence Virtuoso.